refactor: pr feedback
